# openapi-schema.yml
---
openapi: 3.0.3
info:
  title: Category API
  description: This API allows CRUD operations on a category
  contact:
    name: Quarkus
    url: https://github.com/quarkusio
  version: "1.0"
tags:
- name: hello
- name: categories
paths:
  /api/categories:
    get:
      tags:
      - categories
      summary: Returns all the categories from the database
      parameters:
      - name: name_filter
        in: query
        description: An optional filter parameter to filter results by name
        required: false
        schema:
          type: string
      responses:
        "200":
          description: Gets all categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'
              examples:
                categories:
                  value:
                  - id: 1
                    name: Luke Skywalker
                    level: 10
                    picture: https://raw.githubusercontent.com/quarkusio/quarkus-magazenn/characterdata/images/luke-skywalker-2563509063968639219.jpg
                    powers: "Uses light sabre, The force"
    put:
      tags:
      - categories
      summary: Completely replace all categories with the passed-in categories
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/Category'
            examples:
              categories:
                value:
                - id: 1
                  name: Luke Skywalker
                  level: 10
                  picture: https://raw.githubusercontent.com/quarkusio/quarkus-magazenn/characterdata/images/luke-skywalker-2563509063968639219.jpg
                  powers: "Uses light sabre, The force"
        required: true
      responses:
        "201":
          description: The URI to retrieve all the created categories
          headers:
            Location:
              style: simple
              schema:
                format: uri
                type: string
        "400":
          description: Invalid categories passed in (or no request body found)
    post:
      tags:
      - categories
      summary: Creates a valid category
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Category'
            examples:
              valid_category:
                value:
                  name: Luke Skywalker
                  level: 10
                  picture: https://raw.githubusercontent.com/quarkusio/quarkus-magazenn/characterdata/images/luke-skywalker-2563509063968639219.jpg
                  powers: "Uses light sabre, The force"
        required: true
      responses:
        "201":
          description: The URI of the created category
          headers:
            Location:
              style: simple
              schema:
                format: uri
                type: string
        "400":
          description: Invalid category passed in (or no request body found)
    delete:
      tags:
      - categories
      summary: Delete all categories
      responses:
        "204":
          description: Deletes all categories
  /api/categories/hello:
    get:
      tags:
      - hello
      summary: Ping hello
      responses:
        "200":
          description: Ping hello
          content:
            text/plain:
              schema:
                type: string
              examples:
                hello_success:
                  value: Hello Category Resource
  /api/categories/random:
    get:
      tags:
      - categories
      summary: Returns a random category
      responses:
        "200":
          description: Gets a random category
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
              examples:
                category:
                  value:
                    id: 1
                    name: Luke Skywalker
                    level: 10
                    picture: https://raw.githubusercontent.com/quarkusio/quarkus-magazenn/characterdata/images/luke-skywalker-2563509063968639219.jpg
                    powers: "Uses light sabre, The force"
        "404":
          description: No category found
  /api/categories/{id}:
    get:
      tags:
      - categories
      summary: Returns a category for a given identifier
      parameters:
      - name: id
        in: path
        required: true
        schema:
          format: int64
          type: integer
      responses:
        "200":
          description: Gets a category for a given id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
              examples:
                category:
                  value:
                    id: 1
                    name: Luke Skywalker
                    level: 10
                    picture: https://raw.githubusercontent.com/quarkusio/quarkus-magazenn/characterdata/images/luke-skywalker-2563509063968639219.jpg
                    powers: "Uses light sabre, The force"
        "404":
          description: The category is not found for a given identifier
    put:
      tags:
      - categories
      summary: Completely updates/replaces an exiting category by replacing it with the
        passed-in category
      parameters:
      - name: id
        in: path
        required: true
        schema:
          format: int64
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Category'
            examples:
              valid_category:
                value:
                  id: 1
                  name: Luke Skywalker
                  level: 10
                  picture: https://raw.githubusercontent.com/quarkusio/quarkus-magazenn/characterdata/images/luke-skywalker-2563509063968639219.jpg
                  powers: "Uses light sabre, The force"
        required: true
      responses:
        "204":
          description: Replaced the category
        "400":
          description: Invalid category passed in (or no request body found)
        "404":
          description: No category found
    delete:
      tags:
      - categories
      summary: Deletes an exiting category
      parameters:
      - name: id
        in: path
        required: true
        schema:
          format: int64
          type: integer
      responses:
        "204":
          description: Deletes a category
    patch:
      tags:
      - categories
      summary: Partially updates an exiting category
      parameters:
      - name: id
        in: path
        required: true
        schema:
          format: int64
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Category'
            examples:
              valid_category:
                value:
                  id: 1
                  name: Luke Skywalker
                  level: 10
                  picture: https://raw.githubusercontent.com/quarkusio/quarkus-magazenn/characterdata/images/luke-skywalker-2563509063968639219.jpg
                  powers: "Uses light sabre, The force"
        required: true
      responses:
        "200":
          description: Updated the category
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
              examples:
                category:
                  value:
                    id: 1
                    name: Luke Skywalker
                    level: 10
                    picture: https://raw.githubusercontent.com/quarkusio/quarkus-magazenn/characterdata/images/luke-skywalker-2563509063968639219.jpg
                    powers: "Uses light sabre, The force"
        "400":
          description: Null category passed in (or no request body found)
        "404":
          description: No category found
components:
  schemas:
    Category:
      required:
      - name
      - level
      type: object
      properties:
        id:
          format: int64
          type: integer
        name:
          maxLength: 50
          minLength: 3
          type: string
        otherName:
          type: string
        level:
          format: int32
          minimum: 0
          exclusiveMinimum: true
          type: integer
        picture:
          type: string
        powers:
          type: string